many changes

+++ src/main/java/com/example/springapi/repositories/CategoryResponsitory.java
@@ -12,8 +12,7 @@ import com.example.springapi.models.Category;
 public interface CategoryResponsitory extends JpaRepository<Category, Long>{
 
     List<Category> findByName(String name);
-
-    Optional<Category> findById(int id);
+    Optional<Category> findById(Long id);
 
 	
 }
